move ContainerStorage to dao.base
