+++ api/dao/containerstorage.py
@@ -1,202 +1,18 @@
 import datetime
 
-import bson.errors
-import bson.objectid
-import pymongo.errors
+import bson
 
 from . import APIStorageException, APIConflictException, APINotFoundException
-from . import consistencychecker
 from . import containerutil
 from . import hierarchy
 from .. import config
-from .. import util
+from ..jobs.gears import validate_gear_config, get_gear
+from ..jobs.jobs import Job
+from .base import ContainerStorage
 
 
 log = config.log
 
-# TODO: Find a better place to put this until OOP where we can just call cont.children
-CHILD_MAP = {
-    'groups':   'projects',
-    'projects': 'sessions',
-    'sessions': 'acquisitions'
-}
-
-# All "containers" are required to return these fields
-# 'All' includes users
-BASE_DEFAULTS = {
-    '_id':      None,
-    'created':  None,
-    'modified': None
-}
-
-# All containers that inherit from 'container' in the DM
-CONTAINER_DEFAULTS = {
-    'permissions':  [],
-    'files':        [],
-    'notes':        [],
-    'tags':         [],
-    'info':         {}
-}
-
-class ContainerStorage(object):
-    """
-    This class provides access to mongodb collection elements (called containers).
-    It is used by ContainerHandler istances for get, create, update and delete operations on containers.
-    Examples: projects, sessions, acquisitions and collections
-    """
-
-    def __init__(self, cont_name, use_object_id = False):
-        self.cont_name = cont_name
-        self.use_object_id = use_object_id
-        self.dbc = config.db[cont_name]
-
-    @classmethod
-    def factory(cls, cont_name, use_object_id=False):
-        """
-        Factory method to aid in the creation of a ContainerStorage instance
-        when cont_name is dynamic.
-        """
-        cont_name = containerutil.pluralize(cont_name)
-        factories = {
-            'default': lambda: cls(cont_name, use_object_id=use_object_id),
-            'groups': GroupStorage,
-            'projects': ProjectStorage,
-            'sessions': SessionStorage,
-            'acquisitions': AcquisitionStorage,
-            'collections': lambda: cls(cont_name, use_object_id=True),
-            'analyses': AnalysisStorage,
-        }
-        storage_factory = factories.get(cont_name, factories['default'])
-        return storage_factory()
-
-    def _fill_default_values(self, cont):
-        if cont:
-            defaults = BASE_DEFAULTS.copy()
-            if self.cont_name not in ['groups', 'users']:
-                defaults.update(CONTAINER_DEFAULTS)
-            defaults.update(cont)
-            cont = defaults
-        return cont
-
-    def get_container(self, _id, projection=None, get_children=False):
-        cont = self.get_el(_id, projection=projection)
-        if cont is None:
-            raise APINotFoundException('Could not find {} {}'.format(self.cont_name, _id))
-        if get_children:
-            children = self.get_children(_id, projection=projection)
-            cont[CHILD_MAP[self.cont_name]] = children
-        return cont
-
-    def get_children(self, _id, projection=None):
-        try:
-            child_name = CHILD_MAP[self.cont_name]
-        except KeyError:
-            raise APINotFoundException('Children cannot be listed from the {0} level'.format(self.cont_name))
-        query = {self.cont_name[:-1]: bson.objectid.ObjectId(_id)}
-        if not projection:
-            projection = {'metadata': 0, 'files.metadata': 0, 'subject': 0}
-        return self.factory(child_name, use_object_id=True).get_all_el(query, None, projection)
-
-    def _from_mongo(self, cont):
-        return cont
-
-    def _to_mongo(self, payload):
-        return payload
-
-    def exec_op(self, action, _id=None, payload=None, query=None, user=None,
-                public=False, projection=None, recursive=False, r_payload=None,  # pylint: disable=unused-argument
-                replace_metadata=False, unset_payload=None):
-        """
-        Generic method to exec a CRUD operation from a REST verb.
-        """
-
-        check = consistencychecker.get_container_storage_checker(action, self.cont_name)
-        data_op = payload or {'_id': _id}
-        check(data_op)
-        if action == 'GET' and _id:
-            return self.get_el(_id, projection=projection, fill_defaults=True)
-        if action == 'GET':
-            return self.get_all_el(query, user, projection, fill_defaults=True)
-        if action == 'DELETE':
-            return self.delete_el(_id)
-        if action == 'PUT':
-            return self.update_el(_id, payload, unset_payload=unset_payload, recursive=recursive, r_payload=r_payload, replace_metadata=replace_metadata)
-        if action == 'POST':
-            return self.create_el(payload)
-        raise ValueError('action should be one of GET, POST, PUT, DELETE')
-
-    def create_el(self, payload):
-        log.debug(payload)
-        payload = self._to_mongo(payload)
-        try:
-            result = self.dbc.insert_one(payload)
-        except pymongo.errors.DuplicateKeyError:
-            raise APIConflictException('Object with id {} already exists.'.format(payload['_id']))
-        return result
-
-    def update_el(self, _id, payload, unset_payload=None, recursive=False, r_payload=None, replace_metadata=False):
-        replace = None
-        if replace_metadata:
-            replace = {}
-            if payload.get('info') is not None:
-                replace['info'] = util.mongo_sanitize_fields(payload.pop('info'))
-            if payload.get('subject') is not None and payload['subject'].get('info') is not None:
-                replace['subject.info'] = util.mongo_sanitize_fields(payload['subject'].pop('info'))
-
-        update = {}
-
-        if payload is not None:
-            payload = self._to_mongo(payload)
-            update['$set'] = util.mongo_dict(payload)
-
-        if unset_payload is not None:
-            update['$unset'] = util.mongo_dict(unset_payload)
-
-        if replace is not None:
-            update['$set'].update(replace)
-
-        if self.use_object_id:
-            try:
-                _id = bson.objectid.ObjectId(_id)
-            except bson.errors.InvalidId as e:
-                raise APIStorageException(e.message)
-        if recursive and r_payload is not None:
-            hierarchy.propagate_changes(self.cont_name, _id, {}, {'$set': util.mongo_dict(r_payload)})
-        return self.dbc.update_one({'_id': _id}, update)
-
-    def delete_el(self, _id):
-        if self.use_object_id:
-            try:
-                _id = bson.objectid.ObjectId(_id)
-            except bson.errors.InvalidId as e:
-                raise APIStorageException(e.message)
-        return self.dbc.delete_one({'_id':_id})
-
-    def get_el(self, _id, projection=None, fill_defaults=False):
-        if self.use_object_id:
-            try:
-                _id = bson.objectid.ObjectId(_id)
-            except bson.errors.InvalidId as e:
-                raise APIStorageException(e.message)
-        cont = self._from_mongo(self.dbc.find_one(_id, projection))
-        if fill_defaults:
-            cont =  self._fill_default_values(cont)
-        return cont
-
-    def get_all_el(self, query, user, projection, fill_defaults=False):
-        if user:
-            if query.get('permissions'):
-                query['$and'] = [{'permissions': {'$elemMatch': user}}, {'permissions': query.pop('permissions')}]
-            else:
-                query['permissions'] = {'$elemMatch': user}
-        log.debug(query)
-        log.debug(projection)
-        results = list(self.dbc.find(query, projection))
-        for cont in results:
-            cont = self._from_mongo(cont)
-            if fill_defaults:
-                cont =  self._fill_default_values(cont)
-        return results
 
 class GroupStorage(ContainerStorage):
 
@@ -437,7 +253,14 @@ class AcquisitionStorage(ContainerStorage):
         return self.get_all_el(query, user, projection)
 
 
+class CollectionStorage(ContainerStorage):
+
+    def __init__(self):
+        super(CollectionStorage, self).__init__('collections', use_object_id=True)
+
+
 class AnalysisStorage(ContainerStorage):
+
     def __init__(self):
         super(AnalysisStorage, self).__init__('analyses', use_object_id=True)
 
@@ -449,7 +272,7 @@ class AnalysisStorage(ContainerStorage):
 
     def get_analyses(self, parent_type, parent_id, inflate_job_info=False):
         parent_type = containerutil.singularize(parent_type)
-        parent_id = bson.objectid.ObjectId(parent_id)
+        parent_id = bson.ObjectId(parent_id)
         analyses = self.get_all_el({'parent.type': parent_type, 'parent.id': parent_id}, None, None)
         if inflate_job_info:
             for analysis in analyses:
@@ -462,9 +285,9 @@ class AnalysisStorage(ContainerStorage):
         defaults = {
             'parent': {
                 'type': containerutil.singularize(cont_name),
-                'id': bson.objectid.ObjectId(cid)
+                'id': bson.ObjectId(cid)
             },
-            '_id': bson.objectid.ObjectId(),
+            '_id': bson.ObjectId(),
             'created': datetime.datetime.utcnow(),
             'modified': datetime.datetime.utcnow(),
             'user': origin.get('id'),
@@ -481,10 +304,6 @@ class AnalysisStorage(ContainerStorage):
         """
         Create and insert job and analysis.
         """
-
-        # TODO fix import cycle - separate analysisstorage module?
-        from ..jobs.gears import validate_gear_config, get_gear
-        from ..jobs.jobs import Job
 
         self.fill_values(analysis, cont_name, cid, origin)
 
@@ -544,9 +363,6 @@ class AnalysisStorage(ContainerStorage):
         If job is in failed state, look for most recent job referencing this analysis
         Update analysis if new job is found
         """
-
-        # TODO fix import cycle - separate analysisstorage module?
-        from ..jobs.jobs import Job
 
         if analysis.get('job') is None:
             return analysis
