+++ api/jobs/gears.py
@@ -13,7 +13,7 @@ import pymongo
 from .. import config
 from .jobs import Job
 from ..dao import APIValidationException
-from ..dao.containerstorage import ContainerStorage
+from ..dao.base import ContainerStorage
 
 log = config.log
 
@@ -58,7 +58,7 @@ def suggest_container(gear, cont_name, cid):
     Given a container reference, suggest files that would work well for each input on a gear.
     """
 
-    root = ContainerStorage.factory(cont_name, True).get_container(cid, projection={'permissions':0}, get_children=True)
+    root = ContainerStorage.factory(cont_name).get_container(cid, projection={'permissions':0}, get_children=True)
     invocation_schema = get_invocation_schema(gear)
 
     schemas = {}
